pyproject.toml
[project]
name = "libelula-inc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.14",
    "jinja2>=3.1.6",
    "pytailwindcss>=0.2.0",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "basedpyright>=1.29.4",
]
